lf commit: lfd: per-repo agent images via .lf/Dockerfile
Build and cache a custom agent image when a repo provides .lf/Dockerfile.
Re-builds when .lf/.docker-stale marker is present. Concurrent builds for
the same repo are serialized. Also bumps Rust toolchain to 1.88 and fixes
INTEGER→BIGINT for Postgres schema_migrations and LIMIT params.

+++ rust/loopflow/src/lfd/store/postgres.rs
@@ -272,7 +272,7 @@ impl RunStore for PostgresStore {
             query.push_str(" ORDER BY started_at DESC");
             if let Some(limit) = limit {
                 query.push_str(&format!(" LIMIT ${}", params.len() + 1));
-                params.push(Box::new(limit as i32));
+                params.push(Box::new(limit as i64));
             }
 
             let params_ref: Vec<&(dyn ToSql + Sync)> =
@@ -723,7 +723,7 @@ impl RunStore for PostgresStore {
             query.push_str(" ORDER BY started_at DESC");
             if let Some(limit) = limit {
                 query.push_str(&format!(" LIMIT ${}", params.len() + 1));
-                params.push(Box::new(limit as i32));
+                params.push(Box::new(limit as i64));
             }
 
             let params_ref: Vec<&(dyn ToSql + Sync)> = params.iter().map(|v| v.as_ref()).collect();
